Bug Scrubbing and Refactoring (#8)
* Move actions/filters to init function

The _construct method should only help instantiate the class. Don't
make it do more than it should.

* Fix bug where regex would return more than form id

* Refactor to fix bug of incorrect return values

* Remove duplicated code

* Use a simple form to trigger form scanning

If the user had clicked the old link and refreshed the page, the
scan would run since it was based on url. Using a form causes the
scan to only run after the form has been submitted since POST vars
don't persist.

* Use correct plugin name

* Give the user more information about scan state

Let the user know when the scan has started and when it has finished.

* Relocate and style the scan button

* Exclude phpcs from checking tests

* Add tests for get_shortcode_id

* Fix bug where updating posts would create excess relations

* Allow for posts to have multiple forms

* Update travis to current suggested configuration

* Have phpcs check all files in /src

* Fix bug causing only some forms to be found

* Reconfigure location table so it has pagination

* Refactor regex to fix issues with malformed shortcodes

* Fix phpcs errors

* Fix more phpcs errors

* Add an error message for invalid shortcodes

If the user has a shortcode that references a form that does not
exist, let them know.

* Update to version 1.2.0

* Reword the plugin description

* Update changelog

* Add newly reworded description

* Fix incorrect shortcode example

* Remove incorrect comment